requirements.txt
anthropic>=0.40.0
flask>=3.0
requests>=2.31
websocket-client>=1.7
APScheduler>=3.10
google-api-python-client>=2.100
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
python-dotenv>=1.0
deepgram-sdk>=3.0.0
elevenlabs>=1.0.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
